% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{count_words}
\alias{count_words}
\title{Count words in a file, saving a list of distinct words with frequencies.}
\usage{
count_words(input_file, output_file)
}
\arguments{
\item{input_file}{path to input file}

\item{output_file}{path to output file}
}
\value{
number of distinct words found
}
\description{
Count words in a file, saving a list of distinct words with frequencies.
}
