% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_data.R
\docType{data}
\name{small_data}
\alias{small_data}
\title{Sample word frequency data.}
\format{A data frame
\describe{
  \item{word}{The word being counted (chr)}
  \item{n}{The number of occurrences (int)}
}}
\usage{
small_data
}
\description{
This small dataset contains word frequencies for tutorial purposes.
}
\keyword{datasets}
