% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{tally_words}
\alias{tally_words}
\title{Count words in a vector.}
\usage{
tally_words(all_words)
}
\arguments{
\item{all_words}{Character vector of words}
}
\value{
tibble with (word, n) pairs
}
\description{
Count words in a vector.
}
